Outfactor populate script
